requirements.txt
Django==2.1.7
django-ckeditor==5.6.1
django-js-asset==1.2.2
django-mdeditor==0.1.13
Pillow==5.4.1
PyMySQL==0.9.3
pytz==2018.9
